move the function in a better module named
