+++ sklearn/metrics/_plot/roc_curve.py
@@ -1,10 +1,11 @@
-from .._base import _get_response
-
 from .. import auc
 from .. import roc_curve
 
 from ...base import is_classifier
-from ...utils import check_matplotlib_support
+from ...utils import (
+    check_matplotlib_support,
+    _get_response,
+)
 from ...utils.validation import _deprecate_positional_args
