+++ sklearn/metrics/_scorer.py
@@ -33,8 +33,6 @@ from . import (r2_score, median_absolute_error, max_error, mean_absolute_error,
                balanced_accuracy_score, explained_variance_score,
                brier_score_loss, jaccard_score, mean_absolute_percentage_error)
 
-from ._base import _get_response
-
 from .cluster import adjusted_rand_score
 from .cluster import homogeneity_score
 from .cluster import completeness_score
@@ -44,6 +42,7 @@ from .cluster import adjusted_mutual_info_score
 from .cluster import normalized_mutual_info_score
 from .cluster import fowlkes_mallows_score
 
+from ..utils import _get_response
 from ..utils.multiclass import type_of_target
 from ..utils.validation import _deprecate_positional_args
 from ..base import is_regressor
